# docs/swagger_agent_video.yaml
openapi: 3.0.0
info:
  title: Агентный уровень (Видеохостинг)
  description: |
    FastAPI-сервис агентного уровня.
    Поддерживает загрузку MP4-видео, разбиение на минутные фрагменты,
    загрузку фрагментов в S3/MinIO и выдачу служебной информации по фрагментам.
  version: 2.0.0

servers:
  - url: http://localhost:8000
    description: Агентный сервис

tags:
  - name: auth
    description: Прямая проверка логина и пароля без сессий и токенов.
  - name: videos
    description: Операции загрузки и обработки видео.
  - name: fragments
    description: Операции запроса информации о фрагментах.

paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Проверить логин и пароль
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Учетные данные подтверждены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /videos:
    post:
      tags: [videos]
      summary: Загрузить видео и подготовить минутные фрагменты
      operationId: uploadVideo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - user_id
                - title
                - file
              properties:
                user_id:
                  type: string
                  description: Идентификатор пользователя.
                  example: "7d0cf373-2334-45bf-b968-7f44f8ef8fe0"
                title:
                  type: string
                  description: Название видео.
                  example: "Network lecture #1"
                description:
                  type: string
                  nullable: true
                  description: Описание видео.
                  example: "Курсовая работа по сетевым технологиям"
                file:
                  type: string
                  format: binary
                  description: Исходный MP4-файл видео.
      responses:
        '201':
          description: Видео успешно обработано и разложено на фрагменты
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoIngestResponse'
        '400':
          description: Некорректный формат входных данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка обработки видео
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /videos/{video_id}/fragments/{fragment_number}:
    get:
      tags: [fragments]
      summary: Запросить служебную информацию о фрагменте
      operationId: requestVideoFragment
      parameters:
        - name: video_id
          in: path
          required: true
          description: Идентификатор видео.
          schema:
            type: string
            example: "e0db2ccb-cfa8-4427-8394-7f4633495e48"
        - name: fragment_number
          in: path
          required: true
          description: Номер фрагмента, начиная с 1.
          schema:
            type: integer
            minimum: 1
            example: 2
        - name: user_id
          in: query
          required: true
          description: Идентификатор пользователя, инициировавшего запрос.
          schema:
            type: string
            example: "7d0cf373-2334-45bf-b968-7f44f8ef8fe0"
      responses:
        '200':
          description: Служебная информация по фрагменту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FragmentAccessResponse'
        '500':
          description: Внутренняя ошибка сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
          description: Логин пользователя.
          example: student01
        password:
          type: string
          minLength: 1
          maxLength: 256
          description: Пароль пользователя.
          example: strong-password-123

    LoginResponse:
      type: object
      required:
        - status
        - user_id
        - username
        - message
      properties:
        status:
          type: string
          example: ok
        user_id:
          type: string
          example: 7d0cf373-2334-45bf-b968-7f44f8ef8fe0
        username:
          type: string
          example: student01
        message:
          type: string
          example: Login confirmed

    VideoFragmentMeta:
      type: object
      required:
        - fragment_number
        - storage_key
        - content_type
      properties:
        fragment_number:
          type: integer
          description: Порядковый номер минутного фрагмента.
          example: 1
        storage_key:
          type: string
          description: Ключ объекта фрагмента в S3/MinIO.
          example: "videos/e0db2ccb-cfa8-4427-8394-7f4633495e48/fragments/0001.mp4"
        content_type:
          type: string
          example: "video/mp4"
        size_bytes:
          type: integer
          nullable: true
          example: 1548896
        duration_seconds:
          type: integer
          nullable: true
          example: 60
        etag:
          type: string
          nullable: true
          example: "3b83ef96387f14655fc854ddc3c6bd57"

    VideoIngestResponse:
      type: object
      required:
        - video_id
        - title
        - original_filename
        - content_type
        - total_fragments
        - status
        - created_at
        - fragments
      properties:
        video_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        original_filename:
          type: string
        content_type:
          type: string
          example: "video/mp4"
        duration_seconds:
          type: integer
          nullable: true
          example: 305
        total_fragments:
          type: integer
          example: 6
        status:
          type: string
          enum: [processing, ready, failed]
          example: ready
        created_at:
          type: string
          format: date-time
        fragments:
          type: array
          items:
            $ref: '#/components/schemas/VideoFragmentMeta'

    FragmentAccessResponse:
      type: object
      required:
        - request_id
        - status
        - video_id
        - fragment_number
        - created_at
      properties:
        request_id:
          type: string
        status:
          type: string
          enum: [accepted, found, not_found, failed]
        video_id:
          type: string
        fragment_number:
          type: integer
        storage_key:
          type: string
          nullable: true
        content_type:
          type: string
          nullable: true
        object_size:
          type: integer
          nullable: true
        etag:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string